+++ src/components/ColorPicker/components/ColorPickerContent/ColorPickerContent.scss
