+++ src/components/index.js
@@ -61,7 +61,7 @@ export { default as TipseenWizard } from "./Tipseen/TipseenWizard";
 export { default as TipseenImage } from "./Tipseen/TipseenImage";
 export { default as ColorPicker } from "./ColorPicker/ColorPicker";
 export { default as DatePicker } from "./DatePicker/DatePicker";
-export { default as ColorPickerContent } from "./ColorPicker/components/ColorPickerContent/ColorPickerContentComponent";
+export { default as ColorPickerContent } from "./ColorPicker/components/ColorPickerContent/ColorPickerContent";
 export { default as ListItem } from "./ListItem/ListItem";
 export { default as ListItemIcon } from "./ListItemIcon/ListItemIcon";
 export { default as ListItemAvatar } from "./ListItemAvatar/ListItemAvatar";
